requirements.txt
arrow==0.13.1
binaryornot==0.4.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cookiecutter==1.6.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.13
django-rest-swagger==2.2.0
djangorestframework==3.9.2
future==0.17.1
idna==2.8
itypes==1.1.0
Jinja2==2.10
jinja2-time==0.2.0
MarkupSafe==1.1.1
openapi-codec==1.3.2
poyo==0.4.2
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
whichcraft==0.5.2
